+++ backend/src/main/java/org/ftclub/cabinet/auth/service/CookieService.java
@@ -7,7 +7,7 @@ import javax.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.ftclub.cabinet.auth.domain.CookieInfo;
 import org.ftclub.cabinet.auth.domain.CookieManager;
-import org.ftclub.cabinet.config.security.CsrfCookieConfig;
+import org.ftclub.cabinet.security.CsrfCookieConfig;
 import org.ftclub.cabinet.dto.TokenDto;
 import org.ftclub.cabinet.jwt.domain.JwtTokenConstants;
 import org.ftclub.cabinet.jwt.domain.JwtTokenProperties;
