+++ backend/src/main/java/org/ftclub/cabinet/security/CsrfCookieConfig.java
@@ -1,4 +1,4 @@
-package org.ftclub.cabinet.config.security;
+package org.ftclub.cabinet.security;
 
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
