+++ backend/src/main/java/org/ftclub/cabinet/security/RoleHierarchyConfig.java
@@ -1,4 +1,4 @@
-package org.ftclub.cabinet.config.security;
+package org.ftclub.cabinet.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -8,6 +8,9 @@ import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl;
 import org.springframework.security.web.FilterInvocation;
 import org.springframework.security.web.access.expression.DefaultWebSecurityExpressionHandler;
 
+/**
+ * Security 인가 검증 시, MASTER 권한을 최상위로 명시합니다.
+ */
 @Configuration
 public class RoleHierarchyConfig {
