+++ backend/src/main/java/org/ftclub/cabinet/security/exception/SecurityDefaultExceptionHandler.java
@@ -1,4 +1,4 @@
-package org.ftclub.cabinet.config.security.exception;
+package org.ftclub.cabinet.security.exception;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
