+++ backend/src/main/java/org/ftclub/cabinet/security/exception/SecurityExceptionHandler.java
@@ -1,4 +1,4 @@
-package org.ftclub.cabinet.config.security.exception;
+package org.ftclub.cabinet.security.exception;
 
 import java.io.IOException;
 import javax.servlet.http.HttpServletResponse;
